Creating Task005

+++ PythonHomework/Homework 001_lesson1/Task003.py
@@ -2,8 +2,8 @@
 # причём X ≠ 0 и Y ≠ 0 и выдаёт номер четверти плоскости, 
 # в которой находится эта точка (или на какой оси она находится).
 
-coordinate_X = int (input ("Введите координату Х: "))
-coordinate_Y = int (input ("Введите координату Y: "))
+coordinate_X = float (input ("Введите координату Х: "))
+coordinate_Y = float (input ("Введите координату Y: "))
 print ("X =", coordinate_X, ",", "Y =", coordinate_Y)
 if coordinate_X >0 and coordinate_Y >0:
     print ("Точка с координатами X =", coordinate_X, ";", "Y =", coordinate_Y, "лежит в первой четверти.")
